Casino1
This is basically the first one.
These are the following rules.
You get  3 lives to begin with. And the you place bets to win bets as long you have enough hearts/life. The difficulty level of the bets can be chosen every time.
The bet basically involves the player choosing a number and the computer choosing another. If the difficulty level is 1, then your number should be within +- 6  and you get a life if you win. When the level is 2 your number should be within +-4 and you win money and for the last level your number should be within +-2.
Choose wisely and play the game.